Replace bare npx denchclaw with Dench Cloud welcome flow
Bare `npx denchclaw` now shows the Dench Cloud banner and a single
'Continue with Dench.com' action that opens dench.com/login and ends
the session. The full setup pipeline remains on `npx denchclaw
bootstrap`. The Dench Cloud banner moved to its own module so both
flows share it, and the README now documents `npx denchclaw
bootstrap`.
